metaRangeSpecies object

Creates an species object in form of an R6 class that stores and handles all the individual parts that define a species.

Value

A <metaRangeSpecies> object.

Public fields

name

name or ID of the species.

processes

The processes that describe how the species interacts with the environment, itself and other species.

traits

The traits of the species.

sim

A reference to the metaRangeSimulation object that the species is part of. Usefull to acces evironmental data or data of other species.

Methods

Public methods

metaRangeSpecies$new()

metaRangeSpecies$print()

Method new()

Creates a new metaRangeSpecies object

Usage

metaRangeSpecies$new(name, sim)

Arguments

name

<string> name or ID of the species.

sim

<metaRangeSimulation> A reference to the metaRangeSimulation object that the species is part of.

Examples

# The following is bad practice, since species should be added to a simulation
# via the add_species method of the simulation object. But for illustration
# purposes:
sim_env <- terra::sds(terra::rast(nrow = 2, ncol = 2))
test_sim <- metaRangeSimulation$new(source_environment = sim_env)
sp <- metaRangeSpecies$new(name = "species_01", sim = test_sim)
sp

Method print()

Prints information about the species to the console

Usage

metaRangeSpecies$print()

Returns

invisible self

Examples

## ------------------------------------------------
## Method `metaRangeSpecies$new`
## ------------------------------------------------

# The following is bad practice, since species should be added to a simulation
# via the add_species method of the simulation object. But for illustration
# purposes:
sim_env <- terra::sds(terra::rast(nrow = 2, ncol = 2))
test_sim <- metaRangeSimulation$new(source_environment = sim_env)
sp <- metaRangeSpecies$new(name = "species_01", sim = test_sim)
sp
#> Species:  species_01 
#> processes: 
#>  list()
#> traits: 
#>  list()