Metabolic scaling

A function to calculate the metabolic scaling of a parameter.

metabolic_scaling(
  normalization_constant,
  scaling_exponent,
  mass,
  temperature,
  E,
  k = 8.617333e-05
)

Arguments

normalization_constant

<numeric> normalization constant.

scaling_exponent

<numeric> allometric scaling exponent of the mass.

mass

<numeric matrix> mean (individual) mass.

temperature

<numeric matrix> temperature in kelvin (K).

E

<numeric> Activation energy in electronvolts (eV).

k

<numeric> Boltzmann's constant (eV / K).

Value

The scaled parameter.

Details

General notes:

Reproduction rate is generally assumed to scale with an exponent of -1/4 and an activation energy of -0.65 eV (important: it's negative).

Carrying capacity is generally assumed to scale with an exponent of -3/4 and an activation energy of 0.65 eV (important: it's positive). But read: (Brown et. al. 2004; Brown & Sibly, 2012) for an in-depth explanation.

Notes on units:

1 electronvolt = 1.602176634 * 10^-19 Joule

Boltzmann constant 1.380649 * 10^-23 Joule/Kelvin

Boltzmann constant in eV/K = 8.617333e-05 = (1.380649 * 10^-23) / (1.602176634 * 10^-19)

References

Brown, J.H., Gillooly, J.F., Allen, A.P., Savage, V.M. & West, G.B. (2004). Toward a Metabolic Theory of Ecology. Ecology, 85, 1771–1789. doi:10.1890/03-9000

Brown, J.H. and Sibly, R.M. (2012). The Metabolic Theory of Ecology and Its Central Equation. In Metabolic Ecology (eds R.M. Sibly, J.H. Brown and A. Kodric-Brown). doi:10.1002/9781119968535.ch2