metaRange: Mechanistic and Metabolic Constrained Species Distribution Model

A platform to build spatially and temporally explicit process-based species distribution models, that can include an arbitrary number of environmental factors, species and processes including metabolic constraints and species interactions. Its focus is simulating populations of one or multiple species in a grid based landscape and studying the meta-population dynamics and emergent patterns that arise from the interaction of species under complex environmental conditions. 'metaRange' provides functions for common ecological processes such as negative exponential, kernel-based dispersal (see Nathan et al. (2012) doi:10.1093/acprof:oso/9780199608898.003.0015 ), calculation of the envionmental suitability based on cardinal values ( Yan & Hunt (1999) doi:10.1006/anbo.1999.0955 ), reproduction in form of an Ricker model (see Ricker (1954) doi:10.1139/f54-039 and Cabral & Schurr (2010) doi:10.1111/j.1466-8238.2009.00492.x ), as well as metabolic scaling based on the metabolic theory of ecology (see Brown et al. (2004) doi:10.1890/03-9000 and Brown & Sibly (2012) doi:10.1002/9781119968535.ch2 ).

References

tba.

See also

Author

Maintainer: Stefan Fallert stefan.fallert@stud-mail.uni-wuerzburg.de (ORCID)

Authors:

Lea Li lea.li@stud-mail.uni-wuerzburg.de

Juliano S. Cabral j.sarmentocabral@bham.ac.uk (ORCID)

Other contributors:

Bavarian Ministry of Science and Arts (bayklif) [funder]

Deutsche Bundesstiftung Umwelt (DBU) [funder]