Ricker reproduction model

An implementation of the "classic" Ricker reproduction model (Ricker, 1954).

Usage

ricker_reproduction_model(abundance, reproduction_rate, carrying_capacity)

Arguments

abundance

<numeric> vector (or matrix) of abundances.

reproduction_rate

<numeric> vector (or matrix) of reproduction rates.

carrying_capacity

<numeric> vector (or matrix) of carrying capacities.

Value

<numeric> vector (or matrix) of abundances.

Details

Equation:

$$abundance_{t+1} = abundance_t \cdot e^{reproduction\_rate \cdot (1 - \frac{abundance_t}{carrying\_capacity})}$$

Note that the input should have an equal size and that the input abundance should be positive for the results to make sense.

References

Cabral, J.S. and Schurr, F.M. (2010) Estimating demographic models for the range dynamics of plant species. Global Ecology and Biogeography, 19, 85--97. doi:10.1111/j.1466-8238.2009.00492.x

Original model: Ricker, W.E. (1954) Stock and recruitment. Journal of the Fisheries Research Board of Canada, 11, 559--623. doi:10.1139/f54-039

Examples

ricker_reproduction_model(
    abundance = 10,
    reproduction_rate = 0.25,
    carrying_capacity = 100
)
#> [1] 12.52323
ricker_reproduction_model(
    abundance = matrix(10, 10, 5),
    reproduction_rate =  matrix(seq(-0.5, 0.5, length.out = 25), 10, 5),
    carrying_capacity =  matrix(100, 10, 5)
)
#>           [,1]      [,2]      [,3]      [,4]     [,5]
#>  [1,] 6.376282  9.277435 13.498588  7.691264 11.19072
#>  [2,] 6.619932  9.631944 14.014396  7.985162 11.61834
#>  [3,] 6.872893 10.000000 14.549914  8.290291 12.06230
#>  [4,] 7.135520 10.382120 15.105895  8.607080 12.52323
#>  [5,] 7.408182 10.778842 15.683122  8.935973 13.00176
#>  [6,] 7.691264 11.190723  6.376282  9.277435 13.49859
#>  [7,] 7.985162 11.618342  6.619932  9.631944 14.01440
#>  [8,] 8.290291 12.062302  6.872893 10.000000 14.54991
#>  [9,] 8.607080 12.523227  7.135520 10.382120 15.10590
#> [10,] 8.935973 13.001765  7.408182 10.778842 15.68312